Add mircoservices 1-user, 2-game, 3-chat
