+++ game/hand/evaluator.go
