+++ game/hand/evaluator_test.go
