+++ game/hand/lookup.go
