--- server/cover.out
@@ -1,1 +1,0 @@
-mode: set
